General - Fixed confidence Json param for map returning a decimal but expecting int

+++ app/src/main/java/com/example/mobieleapp/data/model/Post.kt
@@ -6,7 +6,7 @@ data class Post (
 ) {
     data class Data(
             val administrative_area: String,
-            val confidence: Int,
+            val confidence: Double,
             val continent: String,
             val country: String,
             val country_code: String,
